requirements.txt
pyrogram
tgcrypto
motor
aiohttp
beautifulsoup4
python-dotenv
pydantic